Updated main.py
To match

+++ main.py
@@ -66,7 +66,7 @@ async def clear_custom_filter(ctx):
     file.close() #close the file
 
     with open("badwords.txt", "a") as f:
-        f.write("if.you.remove.this.line.with.nothing.in.here.the.bot.will.break!") #write stuff to the file. If file is empty the bot will delete everything sent on Discord.
+        f.write("eiufewhiuwefhuwefhwfhueiwfehihewfiwefiuwefiuhwefiuh") #write stuff to the file. If file is empty the bot will delete everything sent on Discord.
     f.close()
 
     await ctx.send(f'Bad word list cleared. Please allow up to 5 seconds while the system restarts.') #tell the author that the bad word list has been cleared.
@@ -89,17 +89,21 @@ async def on_message(message):
     message.content = message.content.lower() #check to see if message is in caps or not.
     message.content = disnake.utils.remove_markdown(message.content) #remove markdown. Exampples: `||test||` " ```test``` "
 
+    if message.content == "eiufewhiuwefhuwefhwfhueiwfehihewfiwefiuwefiuhwefiuh":
+        return
+
+    else:
 
 
-    for word in blacklist: #get each bad word
-        if message.author.bot == True: #check to see if author is a bot. if you want the bot to detect bad words on another bot, remove this line
-            return #return the function of seeing if author is a bot. if you want the bot to detect bad words on another bot, remove this line
-        if message.content.count(word) > 0: #check the content count once.
-            await message.delete() #delete the bad word
-            embed = disnake.Embed(title="Message Deleted", description= f"Bad word blocked by {message.author.mention}", colour=disnake.Color.random()) #send an embed
-            embed.add_field(name = f"User ID:", value = f"{message.author.id}")
-            embed.timestamp = disnake.utils.utcnow()
-            await message.channel.send(embed=embed) #send embed to channel
+        for word in blacklist: #get each bad word
+            if message.author.bot == True: #check to see if author is a bot. if you want the bot to detect bad words on another bot, remove this line
+                return #return the function of seeing if author is a bot. if you want the bot to detect bad words on another bot, remove this line
+            if message.content.count(word) > 0: #check the content count once.
+                await message.delete() #delete the bad word
+                embed = disnake.Embed(title="Message Deleted", description= f"Bad word blocked by {message.author.mention}", colour=disnake.Color.random()) #send an embed
+                embed.add_field(name = f"User ID:", value = f"{message.author.id}")
+                embed.timestamp = disnake.utils.utcnow()
+                await message.channel.send(embed=embed) #send embed to channel
 
 
 #-------------------base-filter------------------------#
@@ -159,16 +163,21 @@ async def on_message(message):
     message.content = message.content.lower()
     message.content = disnake.utils.remove_markdown(message.content)
 
+    if message.content == "eiufewhiuwefhuwefhwfhueiwfehihewfiwefiuwefiuhwefiuh":
+        return
 
-    for word in blacklist2:
-        if message.author.bot == True:
-            return
-        if message.content.count(word) > 0:
-            await message.delete()
-            embed = disnake.Embed(title="Message Deleted", description= f"Bad word blocked by {message.author.mention}", colour=disnake.Color.random()) #send an embed
-            embed.add_field(name = f"User ID:", value = f"{message.author.id}")
-            embed.timestamp = disnake.utils.utcnow()
-            await message.channel.send(embed=embed)
+    else:
+
+
+        for word in blacklist: #get each bad word
+            if message.author.bot == True: #check to see if author is a bot. if you want the bot to detect bad words on another bot, remove this line
+                return #return the function of seeing if author is a bot. if you want the bot to detect bad words on another bot, remove this line
+            if message.content.count(word) > 0: #check the content count once.
+                await message.delete() #delete the bad word
+                embed = disnake.Embed(title="Message Deleted", description= f"Bad word blocked by {message.author.mention}", colour=disnake.Color.random()) #send an embed
+                embed.add_field(name = f"User ID:", value = f"{message.author.id}")
+                embed.timestamp = disnake.utils.utcnow()
+                await message.channel.send(embed=embed) #send embed to channel
 
 
 bot.run ('') #insert the bot token
